% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{ACS}
\alias{ACS}
\title{American Community Survey}
\format{A dataset with 1000 observations on the following 9 variables.
\itemize{
\item{\code{Sex}} { a factor with levels \code{Female} and \code{Male} }
\item{\code{Age}} {age in years }
\item{\code{Married}} {a factor with levels \code{Married} and \code{Not Married} }
\item{\code{Income}}
{Wages and salary for the past 12 months (in \\$1,000's) }
\item{\code{HoursWk}}
{Hours of work per week }
\item{\code{Race}} { a factor with levels \code{asian}, \code{black},
\code{white}, or \code{other} }
\item{\code{USCitizen}} {a factor wit two levels (\code{Citizen} and
\code{Noncitizen}).}
\item{\code{HealthInsurance}} {a factor with two levels (\code{Insured} and
\code{Uninsured}}
\item{\code{Language}} { a factor with two levels (\code{English} and \code{Other})}
}}
\source{
The full public dataset can be downloaded at
\url{http://www.census.gov/acs/www/data_documentation/pums_data/}, and the
full list of variables are at
\url{http://www.census.gov/acs/www/Downloads/datadocumentation/pums/DataDict/PUMSDataDict10.pdf}.
}
\description{
Data from a sample of individuals in the American Community Survey
}
\details{
The American Community Survey, administered by the US Census Bureau, is
given every year to a random sample of about 3.5 million households (about
3\% of all US households).  Data on a random sample of 1\% of all US
residents are made public (after ensuring anonymity), and we have selected a
random sub-sample of n = 1000 from the 2010 data for this dataset.
}
\keyword{datasets}

