% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{APMultipleChoice}
\alias{APMultipleChoice}
\title{AP Multiple Choice}
\format{A dataset with 400 observations on the following variable.
\itemize{
\code{C}, \code{D}, or \code{E} }}
\source{
Sample exams from several disciplines at
\url{http://apcentral.collegeboard.com}
}
\description{
Correct responses on Advanced Placement multiple choice exams
}
\details{
Correct responses from multiple choice sections for a sample of released
Advanced Placement exams
}
\keyword{datasets}

