% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{April14Temps}
\alias{April14Temps}
\title{April 14th Temperatures}
\format{A data frame with 16 observations on the following 3 variables.
\itemize{
   \item{\code{Year}} {1995-2010}

   \item{\code{DesMoines}} {Temperature in Des Moines (degrees F)}

   \item{\code{SanFrancisco}} {Temperature in San Francisco (degrees F)}
   }}
\source{
The University of Dayton Average Daily Temperature Archive at
\url{http://academic.udayton.edu/kissock/http/Weather/citylistUS.htm}
}
\description{
Temperatures in Des Moines, IA and San Francisco, CA on April 14th
}
\details{
Average temperature for the day of April 14th in each of 16 years from
1995-2010
}
\examples{
data(April14Temps)
}
\keyword{datasets}

