% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{AtmosphericCO2}
\alias{AtmosphericCO2}
\title{Atmospheric carbon dioxide levels by year}
\format{A dataset with 11 observations on the following 2 variables.
\itemize{ 
\item{\code{Year}} {Every five years from 1960-2010}
\item{\code{CO2}} {Carbon dioxide level in parts per million}
}}
\source{
Dr. Pieter Tans, NOAA/ESRL (www.esrl.noaa.gov/gmd/ccgg/trends/).
Values recorded at the Mauna Loa Observatory in Hawaii.
}
\description{
Carbon dioxide levels in the atmosphere over a 50 year span from 1960-2010.
}
\keyword{datasets}

