% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{BaseballHits}
\alias{BaseballHits}
\title{Baseball Hits}
\format{A dataset with 30 observations on the following 14 variables.
\itemize{
  \item{\code{Team}} {Name of baseball team}
  \item{\code{League}} {Either \code{AL} or \code{NL}}
  \item{\code{Wins}} {Number of wins for the season}
  \item{\code{Runs}} {Number of runs scored}
  \item{\code{Hits}} {Number of hits}
  \item{\code{Doubles}} {Number of doubles}
  \item{\code{Triples}} {Number of triples}
  \item{\code{HomeRuns}} {Number of home runs}
  \item{\code{RBI}} {Number of runs batted in}
  \item{\code{StolenBases}} {Number of stolen bases}
  \item{\code{CaughtStealing}} {Number of times caught stealing}
  \item{\code{Walks}} {Number of walks}
  \item{\code{Strikeouts}} {Number of stikeouts}
  \item{\code{BattingAvg}} {Team batting average}
  }}
\source{
\url{http://www.baseball-reference.com/leagues/MLB/2011-standard-batting.shtml}
}
\description{
Number of hits and wins for Major League Baseball teams
}
\details{
Data from the 2010 Major League Baseball regular season.
}
\keyword{datasets}

