% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{BaseballTimes}
\alias{BaseballTimes}
\title{Baseball Game Times}
\format{A dataset with 30 observations on the following 9 variables.
\itemize{
  \item{\code{Away} } { Away team name}
  \item{\code{Home}} { Home team name}
  \item{\code{Runs}} { Total runs scored (both teams)}
  \item{\code{Margin}} { Margin of victory}
  \item{\code{Hits}} { Total number of hits (both teams)}
  \item{\code{Errors}} { Total number of errors (both teams)}
\item{\code{Pitchers}} { Total number of pitchers used (both teams)}
\item{\code{Walks}} { Total number of walks (both teams)}
  \item{\code{Time}} { Elapsed time for game (in minutes) }
  }}
\source{
\url{http://www.baseball-reference.com/boxes/2011.shtml}
}
\description{
Information for a sample of 30 Major League Baseball games played during the
2011 season.
}
\details{
Data from a sample of boxscores for Major League Baseball games played in
August 2011.
}
\keyword{datasets}

