% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{Benford}
\alias{Benford}
\title{Benford data}
\format{A dataset with 9 observations on the following 4 variables.
\itemize{
\item{\code{Digit}} {Leading digit (1-9)}
\item{\code{BenfordP}} {Expected proportion according to Benford's law}
\item{\code{Address}} {Frequency as a first digit in an address}
\item{\code{Invoices}} {Frequency as the first digit in invoice amounts}
}}
\source{
Thanks to Prof. Richard Cleary for providing the data
}
\description{
Two examples to test Benford's Law
}
\details{
Leading digits from 1188 addresses sampled from a phone book and 7273
amounts from invoices sampled at a company.
}
\keyword{datasets}

