% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{BikeCommute}
\alias{BikeCommute}
\title{Bike Commute}
\format{A dataset with 56 observations on the following 9 variables.
\itemize{
\item{\code{Bike}} {Type of material (\code{Carbon} or \code{Steel})}
\item{\code{Date}} {Date of the bike commute as a string}
\item{\code{DMY}}  {Date of the bike commute as a date object}
\item{\code{Distance}} {Length of commute (in miles)}
\item{\code{Time}} {Total commute time (hours:minutes:seconds)}
\item{\code{Minutes}} {Time converted to minutes}
\item{\code{AvgSpeed}} {Average speed during the ride (miles per hour)}
\item{\code{TopSpeed}} {Maximum speed (miles per hour)}
\item{\code{Seconds}} {Time converted to seconds}
\item{\code{MonthStr}} {Month}
\code{MonthNum} {Numeric value of month}
\item{\code{Month}} {Month coded as one of \code{1Jan},
\code{2Feb}, \code{3Mar}, \code{4Apr}, \code{5May},
\code{6June}, or \code{7July}},
}}
\source{
Thanks to Dr. Groves for providing his data.
}
\description{
Commute times for two kinds of bicycle
}
\details{
Data from a personal experiment to compare commuting time based on a
randomized selection between two bicycles made of different materials.
}
\references{
"Bicycle weight and commuting time: randomised trial,"
in British Medical Journal, BMJ 2010;341:c6801.
}
\keyword{datasets}

