% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{BodyFat}
\alias{BodyFat}
\title{Body Measurements}
\format{A data frame with 100 observations on the following 10 variables.
\itemize{
   \item{\code{Bodyfat}} {a numeric vector}
   \item{\code{Age}} {a numeric vector}
   \item{\code{Weight}} {a numeric vector}
   \item{\code{Height}} {a numeric vector}
   \item{\code{Neck}} {a numeric vector}
   \item{\code{Chest}} {a numeric vector}
   \item{\code{Abdomen}} {a numeric vector}
   \item{\code{Ankle}} {a numeric vector}
   \item{\code{Biceps}} {a numeric vector}
   \item{\code{Wrist}} {a numeric vector} }}
\description{
Body Measurements
}
\examples{
data(BodyFat)
}
\keyword{datasets}

