% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{BodyTemp50}
\alias{BodyTemp50}
\title{Body Temperatures}
\format{A dataset with 50 observations on the following 3 variables.
\itemize{
\item{\code{BodyTemp}} {Body temperatures in degrees F}
\item{\code{Pulse}} {Pulse rates (beats per minute)}
\item{\code{Sex}} {\code{Female} or \code{Male}}
}}
\source{
Shoemaker, "What's Normal: Temperature, Gender and Heartrate",
Journal of Statistics Education, Vol. 4, No. 2 (1996)
}
\description{
Sample of 50 body temperatures
}
\details{
Body temperatures and pulse rates for a sample of 50 healthy adults.
}
\keyword{datasets}

