% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{BootAtlantaCorr}
\alias{BootAtlantaCorr}
\title{Bootstrap Correlations for Atlanta Commutes}
\format{A data frame with 1000 observations on the following variable.
\itemize{ 
   \item{\code{CorrTimeDist}} {Correlation between Time and Distance
for a bootstrap sample of Atlanta commuters} }}
\source{
Computer simulation
}
\description{
Boostrap correlations between Time and Distance for 500 commuters in Atlanta
}
\details{
Correlations for bootstrap samples of Time vs. Distance for the data on
Atlanta commuters in CommuteAtlanta.
}
\examples{

data(BootAtlantaCorr)

}
\keyword{datasets}

