% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{CaffeineTaps}
\alias{CaffeineTaps}
\title{Caffeine Taps}
\format{A dataset with 20 observations on the following 2 variables.
\itemize{
\item{\code{Taps}} {Number of finger taps in one minute}
\item{\code{Group}} {Treatment with levels \code{Caffeine}
and \code{No Caffeine}}
\item{\code{Caffeine}} {a recoding of \code{Group} with levels \code{"Yes"} and \code{"No"}}
}}
\source{
Hand, Daly, Lund, McConway and Ostrowski, Handbook of Small Data
Sets, Chapman and Hall, London (1994), pp. 40
}
\description{
Finger tap rates with and without caffeine
}
\details{
Results from a double-blind experiment where a sample of male college
students to tap their fingers at a rapid rate. The sample was then divided
at random into two groups of ten students each.  Each student drank the
equivalent of about two cups of coffee, which included about 200 mg of
caffeine for the students in one group but was decaffeinated coffee for the
second group. After a two hour period, each student was tested to measure
finger tapping rate (taps per minute). The goal of the experiment was to
determine whether caffeine produces an increase in the average tap rate.
}
\keyword{datasets}

