% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{CAOSExam}
\alias{CAOSExam}
\title{CAOS Exam Scores}
\format{A datset with 10 observations on the following 3 variables.
\itemize{
\item{\code{Student}} {ID code for student}
\item{\code{Pretest}} {CAOS Pretest score}
\item{\code{Posttest}} {CAOS Posttest score}
}}
\source{
A sample of 10 students from an introductory statisics course.
}
\description{
Scores on a pre-test and post-test of basic statistics concepts
}
\details{
The CAOS (Comprehensive Assessment of Outcomes in First Statistics Course)
exam is designed to measure comprehension basic statistical ideas in an
introductory statistics course.  This dataset has scores for ten students
who took the CAOS pre-test at the start of a course and the post-test during
the course itself.  Each exam consists of 40 multiple choice questions and
the score is the percentage correct.
}
\references{
Find out more about the CAOS exam at
\url{http://app.gen.umn.edu/artist/caos.html}
}
\keyword{datasets}

