% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{CarbonDioxide}
\alias{CarbonDioxide}
\title{CO2}
\format{A dataset with 11 observations on the following 2 variables.
\itemize{
\item{\code{Year}} {Every five years from 1960-2010}
\item{\code{C02}} {Carbon dioxide level in parts per million}
}}
\source{
Dr. Pieter Tans, NOAA/ESRL (www.esrl.noaa.gov/gmd/ccgg/trends/).
Values recorded at the Mauna Loa Observatory in Hawaii.
}
\description{
Atmospheric carbon dioxide levels by year
}
\details{
Carbon dioxide levels in the atmosphere over a 50 year span from 1960-2010.
}
\keyword{datasets}

