% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{Cereal}
\alias{Cereal}
\title{Breakfast Cereals}
\format{A dataset with 30 observations on the following 10 variables.
\itemize{
\item{\code{Name}} {Brand name of cereal}
   \item{ \code{Company}} {Manufacturer coded as
			\code{G}=General Mills, \code{K}=Kellog's or \code{Q}=Quaker}
   \item{\code{Serving}} {Serving size (in cups)}
   \item{\code{Calories}} {Calories (per cup)}
   \item{ \code{Fat}} {Fat (grams per cup)}
   \item{ \code{Sodium}} {Sodium (mg per cup)}
   \item{ \code{Carbs}} {Carbohydrates (grams per cup)}
   \item{ \code{Fiber}} {Dietary Fiber (grams per  cup)}
   \item{ \code{Sugars}} {Sugars (grams per cup)}
   \item{ \code{Protein}} {Protein (grams per cup)}
}}
\source{
Cereal data obtained from nutrition labels at
\url{http://www.nutritionresource.com/foodcomp2.cfm?id=0800}
}
\description{
Nutrition information for a sample of 30 breakfast cereals
}
\details{
Nutrition contents for a sample of breakfast cererals, derived from
nutrition lables.  Values are per cup of cereal (rather than per serving).
}
\keyword{datasets}

