% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{CocaineTreatment}
\alias{CocaineTreatment}
\title{Coacaine Treatment}
\format{A dataset with 72 observations on the following 2 variables.
\itemize{
 \item{\code{Drug}} {Treatment drug: \code{Desipramine},
			\code{Lithium}, or \code{Placebo}}
	\item{\code{Relapse}} { Did the patient relapse? \code{no} or \code{yes}}
}}
\source{
Gawin, F., et.al., "Desipramine Facilitation of Initial Cocaine
Abstinence", Archives of General Psychiatry, 1989; 46(2): 117 - 121.
}
\description{
Relapse/no relapse responses to three different treatments for cocaine
addiction
}
\details{
Data from an experiment to investigate the effectiveness of the two drugs,
desipramine and lithium, in the treatment of cocaine addiction. Subjects
(cocaine addicts seeking treatment) were randomly assigned to take one of
the treatment drugs or a placebo.  The response variable is whether or not
the subject relapsed (went back to using cocaine) after the treatment.
}
\keyword{datasets}

