% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{ColaCalcium}
\alias{ColaCalcium}
\title{Cola Calcium}
\format{A dataset with 16 observations on the following 2 variables.
\itemize{
 	\item{\code{Drink}} {Type of drink:
			\code{Diet cola} or \code{Water}}
		\item{\code{Calcium}} {Amount of calcium excreted (in mg.)}
}}
\source{
Larson, Amin, Olsen, and Poth, Effect of Diet Cola on Urine Calcium
Excretion, Endocrine Reviews, 31[3]: S1070, June 2010. These data are
recreated from the published summary statistics, and are estimates of the
actual data.
}
\description{
Calcium excretion with diet cola and water
}
\details{
A sample of 16 healthy women aged 18 - 40 were randomly assigned to drink 24
ounces of either diet cola or water. Their urine was collected for three
hours after ingestion of the beverage and calcium excretion (in mg.) was
measured . The researchers were investigating whether diet cola leaches
calcium out of the system, which would increase the amount of calcium in the
urine for diet cola drinkers.
}
\keyword{datasets}

