% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{CommuteAtlanta}
\alias{CommuteAtlanta}
\title{Commute Atlanta}
\format{A data frame with 500 observations on the following 5 variables.
\itemize{ St. Louis }}
\source{
Sample chosen using DataFerret at
\url{http://www.thedataweb.org/index.html}.
}
\description{
Commute times and distance for a sample of 500 people in Atlanta
}
\details{
Data from the US Census Bureau's American Housing Survey (AHS) which
contains information about housing and living conditions for samples from
certain metropolitan areas.  These data were extracted respondents in the
Atlanta metropolitan area. They include only cases where the respondent
worked somewhere other than home.  Values show the time (in minutes) and
distance (in miles) that respondents typically traveled on their commute to
work each day as well as age and sex.
}
\examples{
data(CommuteAtlanta)
}
\keyword{datasets}

