% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{CommuteStLouis}
\alias{CommuteStLouis}
\title{Comute times in St. Louis}
\format{A dataset with 500 observations on the following 5 variables.
\itemize{
  	\item{\code{City}} {\code{St. Louis}}
   	\item{\code{Age}} {Age of the respondent (in years)}
		\item{\code{Distance}} {Commute distance (in miles)}
		\item{\code{Time}} {Commute time (in minutes)}
		\item{\code{Sex}} {\code{F} or \code{M}}
}}
\source{
Sample chosen using DataFerret at
\url{http://www.thedataweb.org/index.html}.
}
\description{
Commute times and distance for a sample of 500 people in St. Louis
}
\details{
Data from the US Census Bureau's American Housing Survey (AHS) which
contains information about housing and living conditions for samples from
certain metropolitan areas.  These data were extracted respondents in the
St. Louis metropolitan area. They include only cases where the respondent
worked somewhere other than home.  Values show the time (in minutes) and
distance (in miles) that respondents typically traveled on their commute to
work each day as well as age and sex.
}
\keyword{datasets}

