% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{CompassionateRats}
\alias{CompassionateRats}
\title{Compassionate Rats}
\format{A dataset with 30 observations on the following 2 variables.
\itemize{
		\item{\code{Sex}} {Sex of the rat: coded as \code{F} or
\code{M}}
		\item{\code{Empathy}} {Freed the trapped rat?  \code{no} or \code{yes}}
}}
\source{
Bartal I.B., Decety J., and Mason P., "Empathy and Pro-Social
Behavior in Rats," Science, 2011; 224(6061):1427-1430.
}
\description{
Would a rat attempt to free a trapped rat?
}
\details{
In a recent study, some rats showed compassion by freeing another trapped
rat, even when chocolate served as a distraction and even when the rats
would then have to share the chocolate with their freed companion.
}
\keyword{datasets}

