% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{CricketChirps}
\alias{CricketChirps}
\title{Cricket Chirps}
\format{A dataset with 7 observations on the following 2 variables.
\itemize{
	\item{\code{Temperature}} {Air temperature in degrees F}
		\item{\code{Chirps}} {Cricket chirp rate (chirps per minute)}
}}
\source{
From E.A Bessey and C.A Bessey, Further Notes on Thermometer
Crickets, American Naturalist, (1898) 32, 263-264.
}
\description{
Cricket chirp rate and temperature
}
\details{
The data were collected by E.A. Bessey and C.A. Bessey who measured chirp
rates for crickets and temperatures during the summer of 1898.
}
\keyword{datasets}

