% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{Digits}
\alias{Digits}
\title{Digit counts}
\format{A data frame with 150 observations on the following 7 variables.
\itemize{ 
   \item{\code{Random}} {Four digit random numbers picked by a
sample of students} 
   \item{\code{RND1}} {First digit}

   \item{\code{RND2}} {Second digit} 
   \item{\code{RND3}} {Third digit}

   \item{\code{RND4}} {Fourth digit} 
   \item{\code{SSN8}} {Eighth digit of social
security number} 
   \item{\code{SSN9}} {Last digit of social security number} }}
\source{
In class-student surveys from several classes.
}
\description{
Digits from social security numbers and student selected random numbers
}
\details{
A sample of students were asked to pick a random four digit number.  The
numbers are given in the dataset, along with separate columns for each of
the four digits.  The data also show the last two digits of each student's
social security number (SSN).
}
\examples{

data(Digits)

}
\keyword{datasets}

