% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{DogOwner}
\alias{DogOwner}
\title{Dog/Owner matches}
\format{A data frame with 25 observations on the following variable.
\itemize{ 
   \item{\code{Match}} {Was the dog correctly paired with it's
owner? \code{no} or \code{yes}} }}
\source{
Roy and Christenfeld, Do Dogs Resemble their Owners?, Psychological
Science, Vol. 15, No. 5, 2004, pp. 361 - 363.
}
\description{
Experiment to match dogs with owners
}
\details{
Pictures were taken of 25 owners and their purebred dogs, selected at random
from dog parks. Study participants were shown a picture of an owner together
with pictures of two dogs (the owner's dog and another random dog from the
study) and asked to choose which dog most resembled the owner. Each
dog-owner pair was viewed by 28 naive undergraduate judges, and the pairing
was deemed "correct" (yes) if the majority of judges (more than 14) chose
the correct dog to go with the owner.
}
\examples{

data(DogOwner)

}
\keyword{datasets}

