% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{EmployedACS}
\alias{EmployedACS}
\title{Employed in American Community Survey}
\format{A dataset with 431 observations on the following 9 variables.
\itemize{
	\item{\code{sex}} {\code{Female} or \code{Male} }
	\item{\code{Sex}} {\code{0} or \code{1} }
	\item{\code{Age}} {Age (years)}
	\item{\code{married}} {\code{0} or \code{1}}
	\item{\code{Married}} {\code{Not Married} or \code{Married}}
	\item{\code{Income}} {Wages and salary for the past 12 months (in \$1,000's)}
	\item{\code{HoursWk}} {Hours of work per week}
	\item{\code{Race}} {\code{asian}, \code{black}, \code{white}, or \code{other}}
	\item{\code{UScitizen}} {\code{Citizen} or \code{Noncitizen}}
	\item{\code{USCitizen}} {\code{0} or \code{1}}
		\item{\code{healthInsurance}} {\code{Insured} or \code{Uninsured}}
		\item{\code{HealthInsurance}} {\code{0} or \code{1}}
	\item{code{language}} {Native language: \code{English} or \code{Other}}
	\item{code{Language}} {Native language: \code{0} or \code{1}}
}}
\source{
The full public dataset can be downloaded at 
\url{http://www.census.gov/acs/www/data_documentation/pums_data/}
and the full list of variables is at 
\url{http://www.census.gov/acs/www/Downloads/data/documentation/pums/DataDict/PUMSDataDict10.pdf}.
}
\description{
Employed individiuals from the American Community Survey (ACS) dataset
}
\details{
This is a subset of the ACS dataset including only 431 individuals who were
employed.


Several variables in this data set are included in two encodings.  (Watch your 
capitalization.)  The lowercase versions have more intuitive codings and can 
be used to interpret the numerical codes.
}
\keyword{datasets}

