% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{ExerciseHours}
\alias{ExerciseHours}
\title{Exercise Hours}
\format{A data frame with 50 observations on the following 7 variables.
\itemize{
   \item{\code{Year}} {Year in school (1=First year,..., 4=Senior)}

   \item{\code{Gender}} { \code{F} or \code{M}}
   \item{\code{Hand}} {Left
(\code{l}) or Right (\code{r}) handed?}
   \item{\code{Exercise}} {Hours of
exercise per week}
   \item{\code{TV}} {Hours of TV viewing per week}

   \item{\code{Pulse}} {Resting pulse rate (beats per minute)}

   \item{\code{Pierces}} {Number of body piercings} }}
\source{
In-class student survey.
}
\description{
Amount of exercise per week for students (and other variables)
}
\details{
Data from an in-class survey of statistics students asking about amount of
exercise, TV viewing, handedness, gender, pulse rate, and number of body
piercings.
}
\examples{
data(ExerciseHours)
}
\keyword{datasets}

