% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{FacebookFriends}
\alias{FacebookFriends}
\title{Facebook Friends}
\format{A dataset with 40 observations on the following 2 variables.
\itemize{
 	\item{\code{GMdensity}} {Normalized z-scores of grey matter
				density in certain brain regions}
		\item{\code{FBfriends}} {Number of friends on Facebook}
}}
\source{
Kanai, R., Bahrami, B., Roylance, R., and Rees, G., "Online social
network size is reflected in human brain structure," Proceedings of the
Royal Society, 7 April 2012; 279(1732): 1327-1334. Data approximated from
information in the article.
}
\description{
Data on number of Facebook friends and grey matter density in brain regions
related to social perception and associative memory.
}
\details{
A recent study in Great Britain examines the relationship between the number
of friends an individual has on Facebook and grey matter density in the
areas of the brain associated with social perception and associative memory.
The study included 40 students at City University London.
}
\keyword{datasets}

