% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{FatMice18}
\alias{FatMice18}
\title{Fat Mice 18}
\format{A dataset with 18 observations on the following 2 variables.
\itemize{
	\item{\code{Light}} {Light treatment: \code{LD}= normal
		light/dark cycle or \code{LL}=bright light at night}
	\item{\code{WgtGain4}} {Weight gain (grams over a four week period)}
}}
\source{
Fonken, L., et. al., "Light at night increases body mass by shifting
time of food intake," Proceedings of the National Academy of Sciences,
October 26, 2010; 107(43): 18664-18669.
}
\description{
Weight gain for mice with different nighttime light conditions
}
\details{
This is a subset of the LightatNight dataset, showing body mass gain in mice
after 4 weeks for two of the treatment conditions: a normal light/dark cycle
(LD) or a bright light on at night (LL).
}
\keyword{datasets}

