% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{FishGills3}
\alias{FishGills3}
\title{FishGills3}
\format{A data frame with 360 observations on the following 2 variables.
\itemize{ 
   \item{\code{Calcium}} {Level of Calcium \code{Low} 0.71 mg.L
\code{Medium} 5.24 mg/L \code{High} 18.24 mg/L}

   \item{\code{GillRate}} {Respiration rate (beats per minute)} }}
\source{
Thanks to Pro. Brad Baldwin for supplying the data
}
\description{
Respiration rate for fish in three levels of calcium.
}
\details{
Fish were randomly assigned to three tanks with different levels (low,
medium and high) levels of calcium.  Respiration rate was measured as number
of gill beats per minute.
}
\examples{

data(FishGills3)
}
\keyword{datasets}

