% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{Flight179}
\alias{Flight179}
\title{Flight times}
\format{A data frame with 36 observations on the following 3 variables.
\itemize{
   \item{\code{Date}} {Date of the flight (5th, 15th and 25th of
each month in 2010) as a factor}
   \item{\code{MDY}} {Date as a date object}
   \item{\code{Flight179}} {Flying time (Boston-SF) in
minutes}
   \item{\code{Flight180}} {Fllying time (SF-Boston) in minutes} }}
\source{
Data collected from the Bureau of Transportation Statistics website
at
\url{http://www.bts.gov/xml/ontimesummarystatistics/src/dstat/OntimeSummaryAirtime.xml}
}
\description{
Flight times for Flight 179 (Boston-SF) and Flight 180 (SF-Boston)
}
\details{
United Airlines Flight 179 is a daily flight from Boston to San Francisco.
Flight 180 goes in the other direction (SF to Boston). The data show the
airborn flying times for each flight on the three dates each month (5th,
15th and 25th) in 2010.
}
\examples{
data(Flight179)
}
\keyword{datasets}

