% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{FloridaLakes}
\alias{FloridaLakes}
\title{Florida Lakes}
\format{A data frame with 53 observations on the following 12 variables.
\itemize{
   \item{\code{ID}} {An identifying number for each lake}

   \item{\code{Lake}} {Name of the lake}

   \item{\code{Alkalinity}} {Concentration of calcium carbonate (in mg/L)}

   \item{\code{pH}} {Acidity}
   \item{\code{Calcium}} {Amount of calcium in
water}
   \item{\code{Chlorophyll}} {Amount of chlorophyll in water}

   \item{\code{AvgMercury}} {Average mercury level for a sample of fish (large
mouth bass) from each lake}
   \item{\code{NumSamples}} {Number of fish sampled
at each lake}
   \item{\code{MinMercury}} {Minimum mercury level in a sampled
fish}
   \item{\code{MaxMercury}} {Maximum mercury level in a sampled fish}

   \item{\code{ThreeYrStdMercury}} {Adjusted mercury level to account for the
age of the fish}
   \item{\code{AgeData}} {Mean age of fish in each sample} }}
\source{
Lange, Royals, and Connor, Transactions of the American Fisheries
Society (1993)
}
\description{
Water quality for a sample of lakes in Florida
}
\details{
This dataset describes characteristics of water and fish samples from 53
Florida lakes.  Some variables (e.g. Alkalinity, pH, and Calcium) reflect
the chemistry of the water samples. Mercury levels were recorded for a
sample of large mouth bass selected at each lake.
}
\examples{
data(FloridaLakes)
}
\keyword{datasets}

