% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{GlobalInternet}
\alias{GlobalInternet}
\title{Global Internet Usage}
\format{A data frame with 9 observations on the following 3 variables.
\itemize{
   \item{\code{Country}} {Name of country}
   \item{\code{PercentFastConnection}} {Percent of internet users with a fast
			connection}
   \item{\code{HoursOnline}} {Average number of hours online in
			February 2011}
}}
\source{
NielsenWire, "Swiss Lead in Speed: Comparing Global Internet
Connections", April 1, 2011
}
\description{
Internet usage for several countries
}
\details{
The Nielsen Company measured connection speeds on home computers in nine
different countries.  Variables include the percent of internet users with a
fast connection (defined as 2Mb/sec or faster) and the average amount of
time spent online, defined as total hours connected to the web from a home
computer during the month of February 2011.
}
\examples{
data(GlobalInternet)
}
\keyword{datasets}

