% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{GPAGender}
\alias{GPAGender}
\title{GPA and Gender}
\format{A dataset with 343 observations on the following 6 variables.
\itemize{
	\item{\code{Exercise}} {Hours of exercise (per week)}
		\item{\code{SAT}} {Combined SAT scores (out of 1600)}
	\item{\code{GPA}} {Grade Point Average (0.00-4.00 scale)}
	\item{\code{Pulse}} {Pulse rate (beats per minute)}
	\item{\code{Piercings}} {Number of body piercings}
		\item{\code{Sex}} {\code{Female} or \code{Male}}
}}
\source{
A first day survey over several different introductory statistics
classes.
}
\description{
Data from a survey of introductory statistics students.
}
\details{
This is a subset of the StudentSurvey dataset where cases with missing
values have been dropped and gender is coded as a 0/1 indicator variable.
}
\keyword{datasets}

