% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{HappyPlanetIndex}
\alias{HappyPlanetIndex}
\title{Happy Planet Index}
\format{A dataset with 143 observations on the following 11 variables.
\itemize{
	   \item{\code{Country}} {Name of country}
    \item{\code{Region}} {Three-digit country code}
    \item{\code{Happiness}} {Score on a 0-10 scale for
          average level of happiness (10 is happiest)}
    \item{\code{LifeExpectancy}} {Average life expectancy (in years)}
    \item{\code{Footprint}} {Ecological footprint -- a measure of the
				(per capita) ecological impact}
    \item{\code{HLY}} {Happy Life Years - combines life expectancy with well-being}
    \item{\code{HPI}} {Happy Planet Index (0-100 scale)}
    \item{\code{HPIRank}} {HPI rank for the country}
    \item{\code{GDPperCapita}} {Gross Domestic Product (per capita)}
    \item{\code{HDI}} {Human Development Index}
    \item{\code{Population}} {Population (in millions)}
}}
\source{
Data downloaded from \url{http://www.happyplanetindex.org/data/}
}
\description{
Measurements related to happiness and well-being for 143 countries.
}
\details{
Data for 143 countries from the Happy Planet Index Project that works to
quantify indicators of happiness, well-being, and ecological footprint at a
country level.
}
\references{
Marks, N., "The Happy Planet Index", www.TED.com/talks, August
29, 2010.
}
\keyword{datasets}

