% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{HockeyPenalties}
\alias{HockeyPenalties}
\title{Hockey Penalties}
\format{A data frame with 30 observations on the following 2 variables.
\itemize{ 
   \item{\code{Team}} {Name of the team}

   \item{\code{PIMperG}} {Average penaly minutes per game} }}
\source{
Data obtained online at www.nhl.com
}
\description{
Penalty minutes (per game) for NHL teams in 2010-11
}
\details{
Data give the average numeber of penalty minutes for each of the 30 National
HOckey League (NHL) teams during the 2010-11 regualar season.
}
\examples{

data(HockeyPenalties)

}
\keyword{datasets}

