% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{HollywoodMovies2011}
\alias{HollywoodMovies2011}
\title{Hollywood Movies in 2011}
\format{A data frame with 136 observations on the following 14 variables.
\itemize{ 
   \item{\code{Movie}} {a factor with many levels}

   \item{\code{LeadStudio}} {a factor with many levels}

   \item{\code{RottenTomatoes}} {a numeric vector}

   \item{\code{AudienceScore}} {a numeric vector} 
   \item{\code{Story}} {a factor
with many levels} 
   \item{\code{Genre}} {a factor with levels \code{Action}
\code{Adventure} \code{Animation} \code{Comedy} \code{Drama} \code{Fantasy}
\code{Horror} \code{Romance} \code{Thriller}}

   \item{\code{TheatersOpenWeek}} {a numeric vector}

   \item{\code{BOAverageOpenWeek}} {a numeric vector}

   \item{\code{DomesticGross}} {a numeric vector} 
   \item{\code{ForeignGross}} {a
numeric vector} 
   \item{\code{WorldGross}} {a numeric vector}

   \item{\code{Budget}} {a numeric vector} 
   \item{\code{Profitability}} {a
numeric vector} 
   \item{\code{OpeningWeekend}} {a numeric vector} }}
\description{
Hollywood Movies in 2011
}
\examples{

data(HollywoodMovies2011)
## maybe str(HollywoodMovies2011) ; plot(HollywoodMovies2011) ...

}
\keyword{datasets}

