% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{HomesForSale}
\alias{HomesForSale}
\title{Home for Sale}
\format{A data frame with 120 observations on the following 5 variables.
\itemize{ 
   \item{\code{State}} {Location of the home: \code{CA} \code{NJ}
\code{NY} \code{PA}} 
   \item{\code{Price}} {Asking price (in \$1,000's)}

   \item{\code{Size}} {Area of all rooms (in 1,000's sq. ft.)}

   \item{\code{Beds}} {Number of bedrooms} 
   \item{\code{Baths}} {Number of
bathrooms} }}
\description{
Data on homes for sale in four states
}
\details{
Data for samples of homes for sale in each state, selected from zillow.com.
}
\examples{

data(HomesForSale)

}
\references{
Data collected from www.zillow.com in 2010.
}
\keyword{datasets}

