% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{HomesForSaleCA}
\alias{HomesForSaleCA}
\title{Home for Sale in California}
\format{A data frame with 30 observations on the following 5 variables.
\itemize{
   \item{\code{State}} {Location of the home: \code{CA}}

   \item{\code{Price}} {Asking price (in \$1,000's)}
   \item{\code{Size}} {Area of
all rooms (in 1,000's sq. ft.)}
   \item{\code{Beds}} {Number of bedrooms}

   \item{\code{Baths}} {Number of bathrooms} }}
\source{
Data collected from www.zillow.com in 2010.
}
\description{
Data for a sample of homes offered for sale in California
}
\details{
This is a subset of the HomesForSale data with just information from homes
in California (CA). Data were collected offerings listed on an online site.
}
\examples{
data(HomesForSaleCA)
}
\keyword{datasets}

