% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{HomesForSaleCanton}
\alias{HomesForSaleCanton}
\title{Homes for sale in Canton, NY}
\format{A data frame with 10 observations on the following variable.
\itemize{
   \item{\code{Price}} {Asking price for the home (in \$1,000's)} }}
\source{
Data collected from www.zillow.com in 2010.
}
\description{
Prices of homes for sale in Canton, NY
}
\details{
Data for samples of homes for sale in Canton, NY, selected from zillow.com.
}
\examples{
data(HomesForSaleCanton)
}
\keyword{datasets}

