% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{HomesForSaleNY}
\alias{HomesForSaleNY}
\title{Home for Sale in New York}
\format{A data frame with 30 observations on the following 5 variables.
\itemize{ 
   \item{\code{State}} {Location of the home: \code{NY}}

   \item{\code{Price}} {Asking price (in \$1,000's)} 
   \item{\code{Size}} {Area of
all rooms (in 1,000's sq. ft.)} 
   \item{\code{Beds}} {Number of bedrooms}

   \item{\code{Baths}} {Number of bathrooms} }}
\source{
Data collected from www.zillow.com in 2010.
}
\description{
Data for a sample of homes offered for sale in New York State
}
\details{
This is a subset of the HomesForSale data with just information from homes
in New York State (NY). Data were collected offerings listed on an online
site.
}
\examples{

data(HomesForSaleNY)

}
\keyword{datasets}

