% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{HoneybeeCircuits}
\alias{HoneybeeCircuits}
\title{Honeybee Circuits}
\format{A dataset with 78 observations on the following 2 variables.
\itemize{
		\item{\code{Circuits}} {Number of waggle dance circuits for a
					returning scout bee}
		\item{\code{Quality}} {Quality of the nest site: \code{High} or \code{Low}}
}}
\source{
Seeley, T., Honeybee Democracy, Princeton University Press,
Princeton, NJ, 2010, p. 128
}
\description{
Number of circuits for honeybee dances and nest quality
}
\details{
When honeybees are looking for a new home, they send out scouts to explore
options. When a scout returns, she does a "waggle dance" with multiple
circuit repetitions to tell the swarm about the option she found. The bees
then decide between the options and pick the best one. Scientists wanted to
find out how honeybees decide which is the best option, so they took a swarm
of honeybees to an island with only two possible options for new homes: one
of very high honeybee quality and one of low quality. They then kept track
of the scouts who visited each option and counted the number of waggle dance
circuits each scout bee did when describing the option.
}
\keyword{datasets}

