% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{HoneybeeWaggle}
\alias{HoneybeeWaggle}
\title{Honeybee Waggle}
\format{A dataset with 7 observations on the following 2 variables.
\itemize{
		\item{\code{Distance}}{Distance to the potential nest site (in meters)}
		\item{\code{Duration}} {Duration of the waggle dance (in seconds)}
}}
\source{
Seeley, T., Honeybee Democracy, Princeton University Press,
Princeton, NJ, 2010, p. 128
}
\description{
Honeybee dance duration and distance to nesting site
}
\details{
When honeybee scouts find a food source or a nice site for a new home, they
communicate the location to the rest of the swarm by doing a "waggle dance."
They point in the direction of the site and dance longer for sites farther
away. The rest of the bees use the duration of the dance to predict distance
to the site.
}
\keyword{datasets}

