% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{HotDogs}
\alias{HotDogs}
\title{Hot Dog Eating Contest}
\format{A dataset with 10 observations on the following 2 variables.
\itemize{
 	\item{\code{Year}} {Year of the contest: 2002-2011}
		\item{\code{HotDogs}} {Winning number of hot dogs consumed }
}}
\source{
Nathan's Famous webste at
\url{http://nathansfamous.com/contest/hall_of_fame}
}
\description{
Winning number of hot dogs consumed in an eating contest
}
\details{
Every Fourth of July, Nathan's Famous in New York City holds a hot dog
eating contest, in which contestants try to eat as many hot dogs (with buns)
as possible in ten minutes.  The winning number of hot dogs are given for
each year from 2002-2011.
}
\keyword{datasets}

