% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ImmuneTea}
\alias{ImmuneTea}
\title{Immune Tea}
\format{A data frame with 21 observations on the following 2 variables.
\itemize{
   \item{\code{InterferonGamma}} {Measure of interferon gamma
production}
   \item{\code{Drink}} {Type of drink: \code{Coffee} or \code{Tea}}
}}
\source{
Adapted from Kamath, et.al., "Antigens in tea-Beverage prime human V
2V2 T cells in vitro and in vivo for memory and non-memory antibacterial
cytokine responses", Proceedings of the National Academy of Sciences, May
13, 2003.
}
\description{
Interferon gamma production and tea drinking
}
\details{
Eleven healthy non-tea-drinking individuals were asked to drink five or six
cups of tea a day, while ten healthy non-tea and non-coffee-drinkers were
asked to drink the same amount of coffee, which has caffeine but not the
L-theanine that is in tea. The groups were randomly assigned. After two
weeks, blood samples were exposed to an antigen and production of interferon
gamma was measured.
}
\examples{
data(ImmuneTea)
}
\keyword{datasets}

