% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{InkjetPrinters}
\alias{InkjetPrinters}
\title{Inkjet Printers}
\format{A dataset with 20 observations on the following 6 variables.
\itemize{
		\item{\code{Model}} {Model name of printer}
		\item{\code{PPM}} {Printing rate (pages per minute) for a
			benchmark set of print jobs}
		\item{\code{PhotoTime}} {Time (in seconds) to print 4x6 color photos}
		\item{\code{Price}} {Typical retail price (in dollars)}
		\item{\code{CostBW}} {Cost per page (in cents) for printing in
			black & white}
		\item{\code{CostColor}} {Cost per page (in cents) for printing in color}
}}
\source{
Inkjet printer reviews found at
\url{http://www.pcmag.com/reviews/printers}, August 2011.
}
\description{
Data from online reviews of inkjet printers
}
\details{
Information from reviews of inkjet printers at PCMag.com in August 2011.
}
\keyword{datasets}

