% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{MalevolentUniformsNFL}
\alias{MalevolentUniformsNFL}
\title{Malevolent Uniforms NFL}
\format{A data frame with 28 observations on the following 7 variables.
\itemize{
   \item{\code{NFLTeam}} {Team name}

   \item{\code{NFL_Malevolence}} {Score reflecting the "malevolence" of a
team's uniform}
   \item{\code{ZPenYds}} {Z-score for penalty yards} }}
\source{
Frank and Gilovich, "The Dark Side of Self- and Social Perception:
Black Uniforms and Aggression in Professional Sports", Journal of
Personality and Social Psychology, Vol. 54, No. 1, 1988, p. 74-85.
}
\description{
Perceived malevolence of uniforms and penalies for National Football League
(NFL) teams
}
\details{
Participants with no knowledge of the teams rated the jerseys on
characteristics such as timid/aggressive, nice/mean and good/bad. The
averages of these responses produced a "malevolence" index with higher
scores signifying impressions of more malevolent uniforms. To measure
aggressiveness, the authors used the amount of penalty yards converted to
z-scores and averaged for each team over the seasons from 1970-1986.
}
\examples{
data(MalevolentUniformsNFL)
}
\keyword{datasets}

