% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{MammalLongevity}
\alias{MammalLongevity}
\title{Mammal Longevity}
\format{A data frame with 40 observations on the following 3 variables.
\itemize{ 
   \item{\code{Animal}} {Species of mammal}

   \item{\code{Gestation}} {Time from fertilization until birth (in days)}

   \item{\code{Longevity}} {Average lifespan (in years)} }}
\source{
2010 World Almanac, pg. 292.
}
\description{
Longevity and gestation period for mammals
}
\details{
Dataset with average lifespan (in years) and typical gestation period (in
days) for 40 different species of mammals.
}
\examples{

data(MammalLongevity)

}
\keyword{datasets}

