% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{ManhattanApartments}
\alias{ManhattanApartments}
\title{Manhattan Apartment Prices}
\format{A dataset with 20 observations on the following variable.
\itemize{
		\item{\code{Rent}} {Montly rent in dollars}
}}
\source{
Apartments advertised on Craig's List at \url{http://newyork.craigslist.org},
July 5, 2011.
}
\description{
Monthly rent for one-bedroom apartments in Manhattan, NY
}
\details{
Monthly rents for a sample of 20 one-bedroom apartments in Manhattan, NY
that were advertised on Craig's List in July, 2011.
}
\keyword{datasets}

