% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{MarriageAges}
\alias{MarriageAges}
\title{Marriage Ages}
\format{A data frame with 100 observations on the following 2 variables.
\itemize{
   \item{\code{Husband}} {Age of husband at marriage}

   \item{\code{Wife}} {Age of wife at marriage} }}
\source{
Thanks to Linda Casserly, St. Lawrence County Clerk's Office
}
\description{
Ages for husbands and wives from marriage licenses
}
\details{
Data from a sample of 100 marriage licences in St. Lawrence County, NY gives
the ages of husbands and wives for newly married couples.
}
\examples{
data(MarriageAges)
}
\keyword{datasets}

