% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{MastersGolf}
\alias{MastersGolf}
\title{Masters Golf Scores}
\format{A dataset with 20 observations on the following 2 variables.
\itemize{
		\item{\code{First}} {First round score (in relation to par)}
		\item{\code{Final}} {Final four round score (in relation to par)}
}}
\source{
2011 Masters tournament results at
\url{http://www.masters.com/en_US/discover/past_winners.html}
}
\description{
Scores from the 2011 Masters golf tournament
}
\details{
Data for a random sample of 20 golfers who made the cut at the 2011 Masters
golf tournament.
}
\keyword{datasets}

