% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{MentalMuscle}
\alias{MentalMuscle}
\title{Mental Muscle}
\format{A dataset with 32 observations on the following 3 variables.
\itemize{
		\item{\code{Action}} {Treatment: \code{Actual} motions or
				\code{Mental} imaging motions}
		\item{\code{PreFatigue}} {Time (in seconds) to  complete motions before fatigue}
		\item{\code{PostFatigue}} {Time (in seconds) to complete motions after fatigue}
}}
\source{
Data approximated from summary statistics in: Demougeot L. and
Papaxanthis C., "Muscle Fatigue Affects Mental Simulation of Action," The
Journal of Neuroscience, July 20, 2011, 31(29):10712-10720.
}
\description{
Comparing actual movements to mental imaging movements
}
\details{
In this study, participants were asked to either perform actual arm pointing
motions or to mentally imagine equivalent arm pointing motions. Participants
then developed muscle fatigue by holding a heavy weight out horizontally as
long as they could. After becoming fatigued, they were asked to repeat the
previous mental or actual motions. Eight participants were assigned to each
group, and the time in seconds to complete the motions was measured before
and after fatigue.
}
\keyword{datasets}

