% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{MindsetMatters}
\alias{MindsetMatters}
\title{Mindset Matters}
\format{A data frame with 75 observations on the following 14 variables.
\itemize{
   \item{\code{Condition}} {Treatment condition: \code{uninformed} or
\code{informed}}
   \item{\code{Cond}} {Treatment condition: \code{0}=uninformed or
\code{1}=informed}
   \item{\code{Age}} {Age (in years)}

   \item{\code{Wt}} {Original weight (in pounds)}
   \item{\code{Wt2}} {Weight
after 4 weeks (in pounds}
   \item{\code{BMI}} {Original body mass index}

   \item{\code{BMI2}} {Body mass index after 4 weeks}

   \item{\code{Fat}} {Original body fat percentage}
   \item{\code{Fat2}} {Body
fat percentage after 4 weeks}
   \item{\code{WHR}} {Original waist to hip
ratio}
   \item{\code{WHR2}} {Waist to hip ratio}
   \item{\code{Syst}} {Original
systolic blood pressure}
   \item{\code{Syst2}} {Systolic blood pressure after
4 weeks}
   \item{\code{Diast}} {Original diastolic blood pressure}

   \item{\code{Diast2}} {Diastolic blood pressure after 4 weeks} }}
\source{
Crum, A.J. and Langer, E.J. (2007). Mind-Set Matters: Exercise and
the Placebo Effect, Psychological Science, 18:165-171. Thanks to the authors
for supplying the data.
}
\description{
Data from a study of perceived exercise with maids
}
\details{
In 2007 a Harvard psychologist recruited 75 female maids working in
different hotels to participate in a study. She informed 41 maids (randomly
chosen) that the work they do satisfies the Surgeon General's
recommendations for an active lifestyle (which is true), giving them
examples for how why their work is good exercise. The other 34 maids were
told nothing (uninformed). Various chacteristics (weight, body mass index,
...) were recorded for each subject at the start of the experiment and again
four weeks later. Maids with missing values for weight change have been
removed.
}
\examples{
data(MindsetMatters)
}
\keyword{datasets}

