% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{MinistersRum}
\alias{MinistersRum}
\title{Ministers and Rum}
\format{A data frame with 17 observations on the following 3 variables.
\itemize{ 
   \item{\code{Year}} {Every five years from 1860-1940}

   \item{\code{Ministers}} {Number of Methodist ministers in New England}

   \item{\code{Rum}} {Number of barrels of rum imported into Boston} }}
\description{
Number of ministers and rum imports to New England between 1860 and 1940
}
\details{
Data every five years from 1860 to 1940 on he number of Methodist ministers
working in New England and the annual rum imports into Boston.
}
\examples{

data(MinistersRum)

}
\keyword{datasets}

