% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{MustangPrice}
\alias{MustangPrice}
\title{Mustang Prices}
\format{A data frame with 25 observations on the following 3 variables.
\itemize{
   \item{\code{Age}} {Age of the car (in years)}

   \item{\code{Miles}} {Mileage on the car (in 1,000's)}

   \item{\code{Price}} {Asking price in (\$1,000's)} }}
\source{
Student project with data collected from autotrader.com in 2008.
}
\description{
Price, age, and mileage for used Mustangs at an internet website
}
\details{
A statistics student, Gabe McBride, was interested in prices for used
Mustang cars being offered for sale on an internet site.  He sampled 25 cars
from the website and recorded the age (in years), mileage (in thousands of
miles) and asking price (in \$1,000's) for each car in his sample.
}
\examples{
data(MustangPrice)
}
\keyword{datasets}

