% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{NBAPlayers2011}
\alias{NBAPlayers2011}
\title{NBA Players data for 2010-11 Season}
\format{A data frame with 176 observations on the following 25 variables.
\itemize{
   \item{\code{Player}} {Name of player}
   \item{\code{Age}} {Age in
years)}
   \item{\code{Team}} {Team name}
   \item{\code{Games}} {Games played
(out of 82)}
   \item{\code{Starts}} {Games started}

   \item{\code{Mins}} {Minutes played}
   \item{\code{MinPerGame}} {Minutes per
game}
   \item{\code{FGMade}} {Field goals made}
   \item{\code{FGAttempt}} {Field
goals attempted}
   \item{\code{FGPct}} {Three-point field goal percentage}

   \item{\code{FG3Made}} {Three-point field goals made}

   \item{\code{FG3Attempt}} {Three-point field goals attempted}

   \item{\code{FG3Pct}} {Field goal percentage}
   \item{\code{FTMade}} {Free
throws made}
   \item{\code{FTAttempt}} {Free throws attempted}

   \item{\code{FTPct}} {Free throw percentage}

   \item{\code{OffRebound}} {Offensive rebounds}

   \item{\code{DefRebound}} {Defensive rebounds}
   \item{\code{Rebounds}} {Total
rebounds}
   \item{\code{Assists}} {Number of assists}

   \item{\code{Steals}} {Number of steals}
   \item{\code{Blocks}} {Number of
blocked shots}
   \item{\code{Turnovers}} {Number of turnovers}

   \item{\code{Fouls}} {Number of personal fouls}
   \item{\code{Points}} {Number
of points scored} }}
\source{
\url{http://www.basketball-reference.com/leagues/NBA_2011_stats.html}
}
\description{
Data from the 2010-2011 regular season for 176 NBA basketball players.
}
\examples{
data(NBAPlayers2011)
}
\keyword{datasets}

