% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{NBAStandings}
\alias{NBAStandings}
\title{NBA 2010-11 Regular Season Standings}
\format{A data frame with 30 observations on the following 5 variables.
\itemize{ 
   \item{\code{Team}} {Team name} 
   \item{\code{Wins}} {Number of wins
in an 82 game regular season} 
   \item{\code{Losses}} {Number of losses}

   \item{\code{WinPct}} {Proportion of games won} 
   \item{\code{PtsFor}} {Average
points scored per game} 
   \item{\code{PtsAgainst}} {Average points allowed per
game} }}
\source{
Data downloaded from
\url{http://www.basketball-reference.com/leagues/NBA_2011_games.html}
}
\description{
Won-Loss record and statistics for NBA Teams
}
\details{
Won-Loss record and regular season statistics for 30 teams in the National
Basketball Association for the 2010-2011 season
}
\examples{

data(NBAStandings)

}
\keyword{datasets}

