% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{NFLScores2011}
\alias{NFLScores2011}
\title{NFL Game Scores in 2011}
\format{A dataset with 256 observations on the following 11 variables.
\itemize{
\item{\code{Week}} {a numeric vector}
\item{\code{HomeTeam}} {Home team name}
\item{\code{AwayTeam}} {Visiting team name}
\item{\code{HomeScore}} {Points scored by the home team}
\item{\code{AwayScore}} {Points scored by the visiting team}
\item{\code{HomeYards}} {Yards gained by the home team}
\item{\code{AwayYards}} {Yards gained by the visiting team}
\item{\code{HomeTO}} {Turnovers lost by the home team}
\item{\code{AwayTO}} {Turnovers lost by the visiting team}
\item{\code{Date}}  {Date of the game (as a character string)}
\item{\code{YDM}}  {Date of the game (as a date object)}
\item{\code{Day}} {Day of the week: \code{Mon}, \code{Sat}, \code{Sun}, or \code{Thu}}
}}
\source{
NFL scores and game statistics found at
\url{http://www.pro-football-reference.com/years/2011/games.htm}.
}
\description{
Results for all NFL games for the 2011 regular season
}
\details{
Data for all 256 regular season games in the National Football League (NFL)
for the 2011 season.
}
\keyword{datasets}

