% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{OlympicMarathon}
\alias{OlympicMarathon}
\title{2008 Olympic Men's Marathon}
\format{A data frame with 76 observations on the following 5 variables.
\itemize{
   \item{\code{Rank}} {Order of finish}
   \item{\code{Athlete}} {Name
of marathoner}
   \item{\code{Nationality}} {Country of marathoner}

   \item{\code{Time}} {Time as H:MM:SS}
   \item{\code{Minutes}} {Time in minutes}
}}
\source{
\url{http://2008olympics.runnersworld.com/2008/08/mens-marathon-results.html}
}
\description{
Times for all finishers in the men's marathon at the 2008 Olympics
}
\details{
Results for all finishers in the 2008 Men's Olympic marathon in Beijing,
China.
}
\examples{
data(OlympicMarathon)
}
\keyword{datasets}

