% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{OttawaSenators}
\alias{OttawaSenators}
\title{Ottaw Senators hockey team}
\format{A data frame with 24 observations on the following 2 variables.
\itemize{ 
   \item{\code{Points}} {Number of points (goals + assists) scored}

   \item{\code{PenMins}} {Number of penalty minutes} }}
\source{
Data obtained from \url{http://senators.nhl.com/club/stats.htm}.
}
\description{
Data for 24 players on the 2009-10 Ottawa Senators
}
\details{
Points scored and penaly minutes for 24 players (excluding goalies) playing
ice hockey for the Ottawa Senators during the 2009-10 NHL regular season.
}
\examples{

data(OttawaSenators)

}
\keyword{datasets}

