% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{PizzaGirl}
\alias{PizzaGirl}
\title{Pizza Girl Tips}
\format{A dataset with 24 observations on the following 2 variables.
\itemize{
		\item{\code{Tip}} {Amount of tip (in dollars)}
		\item{\code{Shift}} {Which of three different shifts}
}}
\source{
Pizza Girl: Statistical Analysis at
\url{http://slice.seriouseats.com/archives/2010/04/statistical-analysis-of-a-pizza-delivery-shift-20100429.html}.
}
\description{
Data on tips for pizza deliveries
}
\details{
"Pizza Girl" collected data on her deliveries and tips over three different
evening shifts.
}
\keyword{datasets}

