% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{QuizPulse10}
\alias{QuizPulse10}
\title{Quiz vs Lecture Pulse Rates}
\format{A data frame with 10 observations on the following 3 variables.
\itemize{
   \item{\code{Student}} {ID number for the student}

   \item{\code{Quiz}} {Pulse rate (beats per minute) during a quiz}

   \item{\code{Lecture}} {Pulse rate (beats per minute) during a lecture} }}
\source{
In-class data collection
}
\description{
Paired data with pulse rates in a lecture and during a quiz for 10 students
}
\details{
Ten students in an introductory statistics class measured their pulse rate
in two settings: in the middle of a regular class lecture and again while
taking a quiz.
}
\examples{
data(QuizPulse10)
}
\keyword{datasets}

