% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{RandomP50N200}
\alias{RandomP50N200}
\title{Simulated proportions}
\format{A data frame with 5000 observations on the following 2 variables.
\itemize{
   \item{\code{Count}} {Number of simulated "yes" responses in 200
trials}
   \item{\code{Phat}} {Sample proportion (Count/200)} }}
\source{
Computer simulation
}
\description{
Counts and proportions for 5000 simulated samples with n=200 and p=0.50
}
\details{
Results from 5000 simulations of samples of size n=200 from a population
with proportoin of "yes" responses at p=0.50.
}
\examples{
data(RandomP50N200)
}
\keyword{datasets}

