% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{RestaurantTips}
\alias{RestaurantTips}
\title{Restaurant Tips}
\format{A data frame with 157 observations on the following 7 variables.
\itemize{
   \item{\code{Bill}} {Size of the bill (in dollars)}

   \item{\code{Tip}} {Size of the tip (in dollars)}
   \item{\code{CreditCard}} {Paid
with a credit card?  \code{No} or \code{Yes}}
   \item{\code{Credit}} {Paid
with a credit card?  \code{n} or \code{y}}
   \item{\code{Guests}} {Number of
peole in the group}
   \item{\code{Day}} {Day of the week: \code{m}=Monday,
\code{t}=Tuesday, \code{w}=Wednesday, \code{th}=Thursday, or \code{f}=Friday
}
   \item{\code{Server}} {Code for waiter/waitress: \code{A}, \code{B}, or
\code{C}}
   \item{\code{PctTip}} {Tip as a percentage of the bill} }}
\source{
Thanks to Tom DeRosa for providing the tipping data.
}
\description{
Tip data from the First Crush Bistro
}
\details{
The owner of a bistro called First Crush in Potsdam, NY was interested in
studying the tipping patterns of his customers.  He collected restaurant
bills over a two week period that he believes provide a good sample of his
customers. The data recorded from 157 bills include the amount of the bill,
size of the tip, percentage tip, number of customers in the group, whether
or not a credit card was used, day of the week, and a coded identity of the
server.
}
\examples{
data(RestaurantTips)
}
\keyword{datasets}

