% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{RockandRoll}
\alias{RockandRoll}
\title{Rock & Roll Hall of fame}
\format{A data frame with 273 observations on the following 4 variables.
\itemize{
   \item{\code{Inductee}} {Name of he group or individual}

   \item{\code{FemaleMembers}} {\code{Yes} if individual or membr of the group
is female, otherwise \code{No}}
   \item{\code{Category}} {Type of indiviudal
or group: \code{Performer}, \code{Non-performer}, \code{Early INfluence},
\code{Lifetime Achievement}, \code{Sideman}}
   \item{\code{People}} {Number of
people in the group} }}
\source{
Rock & Roll Hall of Fame website,
\url{http://rockhall.com/inductees/alphabetical/}
}
\description{
Groups and Individuals in the Rock and Roll hall of Fame
}
\details{
All inductees of the Rock & Roll Hall of Fame as of 2012.
}
\examples{
data(RockandRoll)
}
\keyword{datasets}

