% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{SalaryGender}
\alias{SalaryGender}
\title{Salary and Gender}
\format{A dataset with 100 observations on the following 4 variables.
\itemize{
\item{\code{Salary}} {Annual salary in \$1,000's}
\item{\code{Gender}} {0=female or 1=male}
\item{\code{Sex}} {\code{Female} or \code{Male}}
\item{\code{Age}} {Age in years}
\item{\code{phD}} {\code{No} or \code{Yes}}
\item{\code{PhD}} {\code{0} or \code{1}}
}}
\source{
\url{http://www.census.gov/acs/www/data_documentation/public_use_microdata_sample/}
}
\description{
Salaries for college teachers
}
\details{
A random sample of college teachers taken from the 2010 American Community
Survey (ACS) 1-year Public Use Microdata Sample (PUMS).
}
\keyword{datasets}

