% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{SkateboardPrices}
\alias{SkateboardPrices}
\title{Skateboard Prices}
\format{A dataset with 20 observations on the following variable.
\itemize{
		\item{\code{Price}} {Selling price in dollars}
}}
\source{
Random sample taken from all skateboards available for sale on eBay
on February 12, 2012.
}
\description{
Prices of skateboards for sale online
}
\details{
Prices for skateboards offered for sale on eBay.
}
\keyword{datasets}

