% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{SleepCaffeine}
\alias{SleepCaffeine}
\title{Sleep Caffeine}
\format{A data frame with 24 observations on the following 2 variables.
\itemize{
   \item{\code{Group}} {Treatment: \code{Caffeine} \code{Sleep}}
   \item{\code{Words}} {Number of words recalled}
}}
\source{
Mednick, Cai, Kanady, and Drummond, "Comparing the benefits of
caffeine, naps and placebo on verbal, motor and perceptual memory",
Behavioural Brain Research, 193 (2008), 79-86.
}
\description{
Experimentn to compare word recall after sleep or caffeine
}
\details{
A random sample of 24 adults were divided equally into two groups and given
a list of 24 words to memorize. During a break, one group takes a 90 minute
nap while another group is given a caffeine pill.  The response variable is
the number of words participants are able to recall following the break.
}
\examples{
data(SleepCaffeine)
}
\keyword{datasets}

