% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{SleepStudy}
\alias{SleepStudy}
\title{Sleep Study}
\format{A dataset with 253 observations on the following 27 variables.
\itemize{
\item{\code{Gender}} {1=male, 0=female}
\item{\code{Sex}} {\code{Female} or \code{Male}}
\item{\code{ClassYear}} {Year in school, 1=first year, ..., 4=senior}
\item{\code{LarkOwl}} {Early riser or night owl?
		\code{Lark}, \code{Neither}, or \code{Owl}}
\item{\code{NumEarlyClass}} {Number of classes per week before 9 am}
\item{\code{earlyClass}} {Indicator for any early classes}
\item{\code{EarlyClass}} {Indicator for any early classes}
\item{\code{GPA}} {Grade point average (0-4 scale)}
\item{\code{ClassesMissed}} {Number of classes missed in a semester}
\item{\code{CognitionZscore}} {Z-score on a test of cognitive skills}
\item{\code{PoorSleepQuality}} {Measure of sleep quality (higher values are poorer sleep)}
\item{\code{DepressionScore}} {Measure of degree of depression}
\item{\code{AnxietyScore}} {Measure of amount of anxiety}
\item{\code{StressScore}} {Measure of amount of stress}
\item{\code{DepressionStatus}} {Coded depression score: \code{normal},
		\code{moderate}, or \code{severe}}
\item{\code{AnxietyStatus}} {Coded anxiety score: 
     \code{normal}, \code{moderate}, or \code{severe}}
\item{\code{Stress}} {Coded stress score: \code{normal} or \code{high}}
\item{\code{DASScore}} {Combined score for depression, anxiety and stress}
\item{\code{Happiness}} {Measure of degree of happiness}
\item{\code{AlcoholUse}} {Self-reported:
		\code{Abstain}, \code{Light}, \code{Moderate}, or \code{Heavy}}
\item{\code{Drinks}} {Number of alcoholic drinks per week}
\item{\code{WeekdayBed}} {Average weekday bedtime (24.0=midnight)}
\item{\code{WeekdayRise}} {Average weekday rise time (8.0=8 am)}
\item{\code{WeekdaySleep}} {Average hours of sleep on weekdays}
\item{\code{WeekendBed}} {Average weekend bedtime (24.0=midnight)}
\item{\code{WeekendRise}} {Average weekend rise time (8.0=8 am)}
\item{\code{WeekendSleep}} {Average weekend bedtime (24.0=midnight)}
\item{\code{AverageSleep}} {Average hours of sleep for all days}
\item{\code{allNighter}} {Had an all-nighter this semester?  \code{Yes} or \code{No}}
\item{\code{AllNighter}} {Had an all-nighter this semester?  \code{0} or \code{1}}
}}
\source{
Onyper, S., Thacher, P., Gilbert, J., Gradess, S., "Class Start
Times, Sleep, and Academic Performance in College: A Path Analysis," April
2012; 29(3): 318-335.  Thanks to the authors for supplying the data.
}
\description{
Data from a study of sleep patterns for college students.
}
\details{
The data were obtained from a sample of students who did skills tests to
measure cognitive function, completed a survey that asked many questions
about attitudes and habits, and kept a sleep diary to record time and
quality of sleep over a two week period.
}
\keyword{datasets}

