% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{SpeedDating}
\alias{SpeedDating}
\title{Speed Dating}
\format{A dataset with 276 observations on the following 22 variables.
\itemize{
\item{\code{DecisionMale}} {Would the male like another date?  \code{Yes} or \code{No}}
\item{\code{DecisionM}} {Would the male like another date?  \code{0} or \code{1}}
   \item{ \code{DecisionFemale}} {Would the female like another date? \code{Yes} or \code{No}}
   \item{ \code{DecisionF}} {Would the female like another date? \code{0} or \code{1}}
   \item{ \code{LikeM}} {How much the male likes his partner (1-10 scale)}
   \item{ \code{LikeF}} {How much the female likes her partner (1-10 scale)}
   \item{ \code{PartnerYesM}} {Male's estimate of chance the female wants
		another date (1-10 scale)}
   \item{ \code{PartnerYesF}} {Female's estimate of chance the male wants another
       date (1-10 scale)}
   \item{ \code{AgeM}} {Male's age (in years)}
   \item{ \code{AgeF}} {Females age (in years)}
   \item{ \code{RaceM}} {Male's race:
		\code{Asian}, \code{Black}, \code{Caucasian}, \code{Latino}, or \code{Other}}
   \item{ \code{RaceF}} {Female's race:
		\code{Asian}, \code{Black}, \code{Caucasian}, \code{Latino}, or \code{Other}}
   \item{\code{AttractiveM}} {Male's rating of female's attractiveness (1-10 scale)}
   \item{ \code{AttractiveF}} {Female's rating of male's attractiveness (1-10 scale)}
   \item{ \code{SincereM}} {Male's rating of female's sincerity (1-10 scale)}
   \item{ \code{SincereF}} {Female's rating of male's sincerity (1-10 scale)}
   \item{ \code{IntelligentM}} {Male's rating of female's intelligence (1-10 scale)}
   \item{ \code{IntelligentF}} {Female's rating of male's intelligence (1-10 scale)}
   \item{ \code{FunM}} {Male's rating of female as fun (1-10 scale)}
   \item{ \code{FunF}} {Female's rating of male as fun (1-10 scale)}
   \item{ \code{AmbitiousM}} {Male's rating of female's ambition (1-10 scale)}
   \item{ \code{AmbitiousF}} {Female's rating of male's ambition (1-10 scale)}
   \item{ \code{SharedInterestsM}} {Male's rating of female's shared interests
            (1-10 scale)}
   \item{ \code{SharedInterestsF}} {Female's rating of male's shared interests
            (1-10 scale)}
}}
\source{
Gelman, A. and Hill, J., Data analysis using regression and
multilevel/hierarchical models, Cambridge University Press: New York, 2007
}
\description{
Data from a sample of four minute speed dates.
}
\details{
Participants were students at Columbia's graduate and professional schools,
recruited by mass email, posted fliers, and fliers handed out by research
assistants. Each participant attended one speed dating session, in which
they met with each participant of the opposite sex for four minutes. Order
and session assignments were randomly determined. After each four minute
"speed date," participants filled out a form rating their date on a scale of
1-10 on various attributes.  Only data from the first date in each session
is recorded here.
}
\keyword{datasets}

