% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{StatGrades}
\alias{StatGrades}
\title{Statistics Exam Grades}
\format{A dataset with 50 observations on the following 3 variables.
\itemize{
		\item{\code{Exam1}} {Score (out of 100 points) on the first exam}
		\item{\code{Exam2}} {Score (out of 100 points) on the first exam}
		\item{\code{Final}} {Score (out of 100 points) on the final exam}
}}
\source{
Random selection of students in an introductory statistics course.
}
\description{
Grades on statistics exams
}
\details{
Exam scores for a sample of students who completed a course using
Statistics: Unlocking the Power of Data as a text.  The dataset contains
scores on Exam 1 (Chapters 1 to 4), Exam 2 (Chapters 5 to 8), and the Final
exam (entire book).
}
\keyword{datasets}

