% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{StatisticsPhD}
\alias{StatisticsPhD}
\title{Statistics PhD Programs}
\format{A data frame with 82 observations on the following 3 variables.
\itemize{ 
   \item{\code{University}} {Name of the school}

   \item{\code{Department}} {Type of department: \code{Biostatistics} or
\code{Statistics}} 
   \item{\code{FTGradEnrollment}} {Full time graduate
student enrollment} }}
\source{
The full list of the 82 Group IV departments was obtained at
\url{http://www.ams.org/profession/data/annual-survey/group_iv}. Data on
enrollment were obtained primarily from Assistantships and Graduate
Fellowships in the Mathematical Sciences, 2009, American Mathematical
Society.
}
\description{
Enrollments in Statistics PhD Programs
}
\details{
Graduate student enrollments in Statistics and Biostatistics departments in
2009. The list does not include combined departments of mathematics and
statistics and does not include departments that did not reply to the AMS
survey.
}
\examples{

data(StatisticsPhD)

}
\keyword{datasets}

