% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{StockChanges}
\alias{StockChanges}
\title{Stock Changes}
\format{A data frame with 50 observations on the following variable.
\itemize{
   \item{\code{SPChange}} {Change in sock price (in dollars)} }}
\description{
Stock price change for a smaple of stocks from the S \& P 500 (August 2-6,
2010)
}
\details{
A random sample of 50 companies from Standard \& Poor's index of 500
companies was selected. The change in the price of the stock (in dollars)
over the 5-day period from August 2 - 6, 2010 was recorded for each company
in the sample.
}
\examples{
data(StockChanges)
}
\keyword{datasets}

