% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{StorySpoilers}
\alias{StorySpoilers}
\title{Story Spoilers}
\format{A dataset with 12 observations on the following 3 variables.
\itemize{
	\item{\code{Story}} {ID for story}
	\item{\code{Spoiler}} {Average (0-10) rating for spoiler version}
	\item{\code{Original}} {Average (0-10) rating for original version}
}}
\source{
Leavitt, J. and Christenfeld, N., "Story Spoilers Don't Spoil
Stories," Psychological Science, published OnlineFirst, August 12, 2011.
}
\description{
Raitngs for stories with and without spoilers
}
\details{
This study investigated whether a story spoiler that gives away the ending
early diminishes suspense and hurts enjoyment?  For twelve different short
stories, the study's authors created a second version in which a spoiler
paragraph at the beginning discussed the story and revealed the outcome.
Each version of the twelve stories was read by at least 30 people and rated
on a 1 to 10 scale to create an overall rating for the story, with higher
ratings indicating greater enjoyment of the story.  Stories 1 to 4 were
ironic twist stories, stories 5 to 8 were mysteries, and stories 9 to 12
were literary stories.
}
\keyword{datasets}

