% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{StressedMice}
\alias{StressedMice}
\title{Stressed Mice}
\format{A data frame with 14 observations on the following 2 variables.
\itemize{
   \item{\code{Time}} {Time spent in darkness (in seconds)}

   \item{\code{Environment}} {Type of environment: \code{Enriched} or
\code{Standard}} }}
\source{
Data approximated from summary statistics in: Lehmann and Herkenham,
"Environmental Enrichment Confers Stress Resiliency to Social Defeat through
an Infralimbic Cortex-Dependent Neuroanatomical Pathway", The Journal of
Neuroscience, April 20, 2011, 31(16):61596173.
}
\description{
Time in darkness for mice in different environments
}
\details{
In the recent study, mice were randomly assigned to either an enriched
environment where there was an exercise wheel available, or a standard
environment with no exercise options.  After three weeks in the specified
environment, for five minutes a day for two weeks, the mice were each
exposed to a "mouse bully" - a mouse who was very strong, aggressive, and
territorial.  one measure of mouse anxiety is amount of time hiding in a
dark compartment, with mice who are more anxious spending more time in
darkness. The amount of time spent in darkness is recorded for each of the
mice.
}
\examples{
data(StressedMice)
}
\keyword{datasets}

