% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{StudentSurvey}
\alias{StudentSurvey}
\title{Student Survey Data}
\format{A data frame with 362 observations on the following 17 variables.
\itemize{ 
   \item{\code{Year}} {Year in school}

   \item{\code{Gender}} {Student's gender: \code{F} or \code{M}}

   \item{\code{Smoke}} {Smoker? \code{No} or \code{Yes}}

   \item{\code{Award}} {Prefered award: \code{Academy} \code{Nobel}
\code{Olympic}} 
   \item{\code{HigherSAT}} {Which SAT is higher?  \code{Math}
or \code{Verbal}} 
   \item{\code{Exercise}} {Hours of exercsie per week}

   \item{\code{TV}} {Hours of TV viewing per week} 
   \item{\code{Height}} {Height
(in inches)} 
   \item{\code{Weight}} {Weight (in pounds)}

   \item{\code{Siblings}} {Number of siblings} 
   \item{\code{BirthOrder}} {Birth
order, 1=oldest} 
   \item{\code{VerbalSAT}} {Verbal SAT score}

   \item{\code{MathSAT}} {Math SAT score} 
   \item{\code{SAT}} {Combined Verbal +
Math SAT} 
   \item{\code{GPA}} {College grade point average}

   \item{\code{Pulse}} {Pulse rate (beats per minute)}

   \item{\code{Piercings}} {Number of body piercings} }}
\source{
In-class student survey
}
\description{
Data from a survey of students in introductory statistics courses
}
\details{
Data from an in-class survey given to introductory statistics students over
several years.
}
\examples{

data(StudentSurvey)

}
\keyword{datasets}

