% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{TenCountries}
\alias{TenCountries}
\title{Ten Countries}
\format{A data frame with 10 observations on the following 4 variables.
\itemize{
   \item{\code{Country}} {a factor with levels \code{Armenia}
\code{Bahamas} \code{Lebanon} \code{Macedonia} \code{North Korea}
\code{Romania} \code{Serbia} \code{Slovenia} \code{Tunisia}
\code{Uzbekistan}}
   \item{\code{Code}} {a factor with levels \code{ARM}
\code{BHS} \code{LBN} \code{MKD} \code{PRK} \code{ROU} \code{SRB} \code{SVN}
\code{TUN} \code{UZB}}
   \item{\code{Area}} {a numeric vector}

   \item{\code{PctRural}} {a numeric vector} }}
\description{
A subset of the ALLCountries data for a random sample of ten countries
}
\examples{
data(TenCountries)
## maybe str(TenCountries) ; plot(TenCountries) ...
}
\keyword{datasets}

