% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{TextbookCosts}
\alias{TextbookCosts}
\title{Textbook Costs}
\format{A data frame with 40 observations on the following 3 variables.
\itemize{
   \item{\code{Field}} {General discipline of the course:
\code{Arts}, \code{Humanities}, \code{NaturalScience}, or
\code{SocialScience}}
   \item{\code{Books}} {Number of books requiired}

   \item{\code{Cost}} {Total cost (in dollars) for required books} }}
\source{
Bookstore online site
}
\description{
Prices for textbooks for different courses
}
\details{
Data are from samples of ten courses in each of four disciplines at a
liberal arts college.  For each course the bookstore's website lists the
required texts(s) and costs. Data were collected for the Fall 2011 semester.
}
\examples{
data(TextbookCosts)
}
\keyword{datasets}

