% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ToenailArsenic}
\alias{ToenailArsenic}
\title{Toenail Arsenic}
\format{A data frame with 19 observations on the following variable.
\itemize{
   \item{\code{Arsenic}} {Level of arsenic found in toenails} }}
\source{
Adapted from Karagas, et.al.,"Toenail Samples as an Indicator of
Drinking Water Arsenic Exposure", Cancer Epidemiology, Biomarkers and
Prevention 1996;5:849-852.
}
\description{
Arsenic in toenails of 19 people using private wells in New Hampshire
}
\details{
Level of arsenic was measured in toenails of 19 subjects from New Hampshire,
all with private wells as their main water source.
}
\examples{
data(ToenailArsenic)
}
\keyword{datasets}

