% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{TrafficFlow}
\alias{TrafficFlow}
\title{Traffic Flow}
\format{A data frame with 24 observations on the following 3 variables.
\itemize{ 
   \item{\code{Timed}} {Delay time (in minutes) for fixed timed
lights} 
   \item{\code{Flexible}} {Delay time (in minutes) for flexible
communicating lights.} 
   \item{\code{Difference}} {Difference (Timed-Flexible)
for each simulation} }}
\source{
Lammer and Helbing, "Self-Stabilizing decentralized signal control
of realistic, saturated network traffic", Santa Fe Institute working paper
\# 10-09-019, September 2010.
}
\description{
Traffic flow times from a simulation with timed and flexible traffic lights
}
\details{
Engineers in Dresden, Germany were looking at ways to improve traffic flow
by enabling traffic lights to communicate information about traffic flow
with nearby traffic lights. The data in show results of one experiment where
they simulated buses moving along a street and recorded the delay time (in
seconds) for both a fixed time and a flexible system of lights. The process
was repeated under both conditions for a sample of 24 simulated scenarios.
}
\examples{

data(TrafficFlow)

}
\keyword{datasets}

