% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{USStates}
\alias{USStates}
\title{US State Data}
\format{A data frame with 50 observations on the following 17 variables.
\itemize{
   \item{\code{State}} {Name of state}

   \item{\code{HouseholdIncome}} {Mean household income (in dollars)}

   \item{\code{IQ}} {Mean IQ score of residents}

   \item{\code{McCainVote}} {Percentage of votes for John McCain in 2008
Presidential election}
   \item{\code{Region}} {Area of the country:
\code{MW}=Midwest, \code{NE}=Northeast, \code{S}=South, or \code{W}=West}

   \item{\code{ObamaMcCain}} {Which 2008 Presidential candidate won state?
\code{M}=McCain or \code{O}=Obama}
   \item{\code{Pres2008}} {Which 2008 Presidential candidate won state?
\code{M}=McCain or \code{O}=Obama}
   \item{\code{Population}} {Number of
residents (in millions)}
   \item{\code{EighthGradeMath}} {a numeric vector}

   \item{\code{HighSchool}} {Percentage of high school graduates}

   \item{\code{GSP}} {Gross State Product (dollars per capita)}

   \item{\code{FiveVegetables}} {Percentage of residents who eat at least five
servings of fruits/vegetables per day}
   \item{\code{Smokers}} {Percentage of
residents who smoke}
   \item{\code{PhysicalActivity}} {Percentage of residents
who have competed in a physical activity in past month}

   \item{\code{Obese}} {Percentage of residents classified as obese}

   \item{\code{College}} {Percentage of residents with college degrees}

   \item{\code{NonWhite}} {Percentage of residents who are not white}

   \item{\code{HeavyDrinkers}} {Percentage of residents who drink heavily} }}
\source{
Various online sources, mostly at www.census.gov
}
\description{
Various data for all 50 US States
}
\examples{
data(USStates)
}
\keyword{datasets}

