% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{WaterStriders}
\alias{WaterStriders}
\title{Water Striders}
\format{A dataset with 10 observations on the following 3 variables.
\itemize{
		\item{\code{AggressiveMale}} {Hyper-aggressive male in group?
			\code{No} or \code{Yes}}
		\item{\code{FemalesHiding}}  {Proportion of time the
				female water striders were in hiding}
		\item{\code{MatingActivity}} {Measure of mean mating activity
				(higher numbers meaning more mating)}
}}
\source{
Sih, A. and Watters, J., "The mix matters: behavioural types and
group dynamics in water striders," Behaviour, 2005; 142(9-10): 1423.
}
\description{
Mating activity for water striders
}
\details{
Water striders are common bugs that skate across the surface of water. Water
striders have different personalities and some of the males are
hyper-aggressive, meaning they jump on and wrestle with any other water
strider near them. Individually, because hyper-aggressive males are much
more active, they tend to have better mating success than more inactive
striders. This study examined the effect they have on a group. Four males
and three females were put in each of ten pools of water. Half of the groups
had a hyper-aggressive male as one of the males and half did not. The
proportion of time females are in hiding was measured for each of the 10
groups, and a measure of mean mating activity was also measured with higher
numbers meaning more mating.
}
\keyword{datasets}

