% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{WaterTaste}
\alias{WaterTaste}
\title{WaterTaste}
\format{A data frame with 100 observations on the following 10 variables.
\itemize{
   \item{\code{Gender}} {Gender of respondent: \code{F}=Female
\code{M}=Male}
   \item{\code{Age}} {Age (in years)}
   \item{\code{Class}} {Year
in school \code{F}=First year \code{J}=Junior \code{O}=Other \code{P}
\code{SO}=Sophomore \code{SR}=Senior}
   \item{\code{UsuallyDrink}} {Usual
source of drinking water: \code{Bottled}, \code{Filtered}, or \code{Tap}}

   \item{\code{FavBotWatBrand}} {Favorite brand of bottled water }

   \item{\code{Preference}} {Order of perference: \code{A}=Sams Choice,
\code{B}=Aquafina, \code{C}=Fiji, and \code{D}=Tap water }

   \item{\code{First}} {Top choice among \code{Aquafina}, \code{Fiji},
\code{SamsChoice}, or \code{Tap}}
   \item{\code{Second}} {Second choice}

   \item{\code{Third}} {Third choice }
   \item{\code{Fourth}} {Fourth choice} }}
\source{
"Water Taste Test Data" by M. Leigh Lunsford and Alix D. Dowling
Finch in the Journal of Statistics Education (Vol 18, No, 1) 2010
}
\description{
Blind taste test to compare brands of bottled water
}
\details{
Result from a blind taste test comparing different four different types of
water (Sam's Choice, Aqufina, Fiji, and tap water).  Participants rank
ordered waters when presented in a random order.
}
\examples{
data(WaterTaste)
}
\references{
\url{http://www.amstat.org/publications/jse/v18n1/lunsford.pdf}
}
\keyword{datasets}

