% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{data}
\name{Wetsuits}
\alias{Wetsuits}
\title{Wetsuits}
\format{A data frame with 12 observations on the following 4 variables.
\itemize{ 
   \item{\code{Wetsuit}} {Maximum swim velocity (m/sec) when wearing
a wetsuit} 
   \item{\code{NoWetsuit}} {Maximum swim velocity (m/sec) when
wearing a regular bathing suit} 
   \item{\code{Gender}} {Gender of swimmer:
\code{F} or \code{M}} 
   \item{\code{Type}} {Type of athlete: \code{swimmer} or
\code{triathlete}} }}
\source{
de Lucas, R.D., Balildan, P., Neiva, C.M., Greco, C.C., Denadai,
B.S. (2000). "The effects of wetsuits on physiological and biomechanical
indices during swimming," Journal of Science and Medicine in Sport, 3 (1):
1-8.
}
\description{
Swim velocity (for 1500 meters) with and withut wearing a wetsuit
}
\details{
A study tested whether wearing wetsuits influences swimming velocity.
Twelve competitive swimmers and triathletes swam 1500m at maximum speed
twice each; once wearing a wetsuit and once wearing a regular bathing suit.
The order of the trials was randomized.  Each time, the maximum velocity in
meters/sec of the swimmer was recorded.
}
\examples{

data(Wetsuits)

}
\keyword{datasets}

