% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{LifeExpectancyVehicles}
\alias{LifeExpectancyVehicles}
\title{Life Expectancy and Vehicle Registrations}
\format{A data frame with 40 observations on the following 3 variables.
\itemize{
   \item{\code{Year}} {a numeric vector}

   \item{\code{LifeExpectancy}} {a numeric vector}
   \item{\code{Vehicles}} {a
numeric vector} }}
\description{
Life Expectancy and Vehicle Registrations
}
\examples{
data(LifeExpectancyVehicles)
## maybe str(LifeExpectancyVehicles) ; plot(LifeExpectancyVehicles) ...
}
\keyword{datasets}

