% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lock5withR-package.R
\docType{package}
\name{Lock5Data-package}
\alias{Lock5Data}
\alias{Lock5Data-package}
\title{Lock5 Datasets}
\description{
Datasets for Statistics: Unlocking the Power of Data \cr by Lock, Lock,
Lock, Lock and Lock
}
\author{
Robin Lock \email{rlock@stlawu.edu}
}

